FastUnitfulOperations

Documentation for FastUnitfulOperations.

FastUnitfulOperations.uproduct

FastUnitfulOperations.uproduct

FastUnitfulOperations.uproduct!

FastUnitfulOperations.uproduct!

FastUnitfulOperations.usort

FastUnitfulOperations.usort!

FastUnitfulOperations.usortperm

FastUnitfulOperations.usortperm!

FastUnitfulOperations.uproduct! — Method

uproduct!(c, a, b)

Fallback to regular multiplication: uproduct!(c, a, b) = mul!(c, a, b)

FastUnitfulOperations.uproduct! — Method

uproduct!(c::AbstractArray, a::AbstractArray{T1}, b::AbstractArray{T2}) where {T1 and/or T2 <:Quantity}

Fast in-place matrix multiplication for Unitful arrays.

This method reinterprets the unitful arrays as arrays with the number's type to make use of optimized matrix multiplication for integers or floats. The method will throw an error if the array contains Unitful elements that do not have a fixed value or unit type.

FastUnitfulOperations.uproduct — Method

uproduct(a, b)
*ᵘ(a, b)

Fallback to regular multiplication: uproduct(a, b) = a * b

FastUnitfulOperations.uproduct — Method

uproduct(a::AbstractArray{T1}, b::AbstractArray{T2}) where {T1 and/or T2 <:Quantity}
*ᵘ(a::AbstractArray{T1}, b::AbstractArray{T2}) where {T1 and/or T2 <:Quantity}

Fast matrix multiplication for Unitful arrays.

This method reinterprets the unitful arrays as arrays with the number's type to make use of optimized matrix multiplication for integers or floats. The method will throw an error if the array contains Unitful elements that do not have a fixed value or unit type.

Type *\^u followed by the tab key to input the operator *ᵘ.

FastUnitfulOperations.usort! — Method

usort!(a::AbstractArray{T}; kwargs...) where {T<:Quantity}
usort!(a; kwargs...)

Sort a unitful array in-place as a reinterpreted array according to the number's type.

Calls Base.sort! on the reinterpreted array. This enables the use of sorting algorithms that only work for integers or floats (e.g. RadixSort from SortingAlgorithms.

The method falls back Base.sort! for other array types.

FastUnitfulOperations.usort — Method

usort(a::AbstractArray{T}; kwargs...) where {T<:Quantity}
usort(a; kwargs...)

Sort a unitful array as a reinterpreted array according to the number's type.

Calls Base.sort on the reinterpreted array. This enables the use of sorting algorithms that only work for integers or floats (e.g. RadixSort from SortingAlgorithms.

The method falls back Base.sort for other array types.

FastUnitfulOperations.usortperm! — Method

usortperm!(ix, a::AbstractArray{T}; kwargs...) where {T<:Quantity}
usortperm!(ix, a; kwargs...)

Return the permutation vector in the preallocated vector ix to sort a unitful array as a reinterpreted array according to the number's type.

Calls Base.sortperm! on the reinterpreted array. This enables the use of sorting algorithms that only work for integers or floats (e.g. RadixSort from SortingAlgorithms.

The method falls back Base.sortperm! for other array types.

FastUnitfulOperations.usortperm — Method

usortperm(a::AbstractArray{T}; kwargs...) where {T<:Quantity}
usortperm(a; kwargs...)

Return the permutation vector to sort a unitful array as a reinterpreted array according to the number's type.

Calls Base.sortperm on the reinterpreted array. This enables the use of sorting algorithms that only work for integers or floats (e.g. RadixSort from SortingAlgorithms.

The method falls back Base.sortperm for other array types.